Create layout of web page
For the first project involving css in TheOdinProject a layout of the
web page was created with html and an image was picked.

5 divs were made for 5 sections including the footer. In each section
subsequent divs were added to get the layout to resemble that of a
generic website. The image selected was put in several locations fitting
the layout given to me.

There is a high chance that changes will be required atfer implementing
css.